requirements.txt
# Core dependencies
numpy==1.24.3
opencv-python==4.8.1
Pillow==10.1.0

# OCR
pytesseract==0.3.10
easyocr==1.7.0

# Document creation
python-docx==0.8.11
reportlab==4.0.7

# Audio processing (from original project)
openai-whisper==20231117
pydub==0.25.1
ffmpeg-python==0.2.0

# Web framework
flask==2.3.3
flask-cors==4.0.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1